Change GetWorkflowEventsRequest.limit and ListJobWorkflowsRequest.instance_events_limit to uint32 type
Test Plan: Unit tests

Reviewers: zhixin, apoorvaj, O521 Project Peloton: Add reviewers

Reviewed By: zhixin, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2709541

+++ pkg/jobmgr/jobsvc/stateless/handler.go
@@ -1344,7 +1344,7 @@ func (h *serviceHandler) ListJobWorkflows(
 func (h *serviceHandler) getInstanceWorkflowEvents(
 	ctx context.Context,
 	updateModel *models.UpdateModel,
-	limit uint64,
+	limit uint32,
 ) ([]*stateless.WorkflowInfoInstanceWorkflowEvents, error) {
 
 	f := func(ctx context.Context, instance_id interface{}) (interface{}, error) {
